backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
supabase==2.3.4

# Auth
python-jose[cryptography]==3.3.0
google-auth==2.27.0
httpx>=0.24,<0.26

# Video Generation
edge-tts==6.1.9
moviepy==1.0.3
Pillow==10.2.0
playwright==1.41.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.3
